refactor: organizing chat function in the same file
