+++ internal/presentation/minecraft_chat.go
@@ -3,9 +3,9 @@ package presentation
 import (
 	"fmt"
 	stdlog "log"
+	"strings"
 
 	"discord-bot-go/internal/connections"
-	"discord-bot-go/internal/log"
 
 	"github.com/bwmarrin/discordgo"
 )
@@ -15,6 +15,11 @@ type MinecraftChatBridge struct {
 	ChannelID string
 
 	Webhook *discordgo.Webhook
+}
+
+type ChatMessage struct {
+	Player  string
+	Message string
 }
 
 func NewMinecraftChatBridge(
@@ -39,7 +44,7 @@ func NewMinecraftChatBridge(
 }
 
 func (b *MinecraftChatBridge) Handle(line string) {
-	msg, ok := log.ParseChatLine(line)
+	msg, ok := ParseChatLine(line)
 	if !ok {
 		return
 	}
@@ -62,4 +67,40 @@ func (b *MinecraftChatBridge) Handle(line string) {
 	if err != nil {
 		stdlog.Printf("Failed to send message via webhook: %v", err)
 	}
+}
+
+func ParseChatLine(line string) (*ChatMessage, bool) {
+	if !strings.Contains(line, "<") || !strings.Contains(line, ">") {
+		return nil, false
+	}
+
+	ignore := []string{
+		"[Rcon]", "Disconnecting VANILLA connection attempt",
+		"rejected vanilla connections", "lost connection", "id=<null>", "legacy=false",
+		"lost connection: Disconnected", "<init>",
+	}
+
+	for _, p := range ignore {
+		if strings.Contains(line, p) {
+			return nil, false
+		}
+	}
+
+	start := strings.Index(line, "<")
+	end := strings.Index(line, ">")
+	if start == -1 || end == -1 || end <= start {
+		return nil, false
+	}
+
+	player := strings.TrimSpace(line[start+1 : end])
+	msg := strings.TrimSpace(line[end+1:])
+
+	if player == "" || msg == "" {
+		return nil, false
+	}
+
+	return &ChatMessage{
+		Player:  player,
+		Message: msg,
+	}, true
 }
